Rename gateway.rs to gateway_config.rs (#1480)
